# docs/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "DrainHostResponse": {
        "additionalProperties": false,
        "properties": {
          "drained_workers": {
            "title": "Drained Workers",
            "type": "integer"
          },
          "host_id": {
            "format": "uuid",
            "title": "Host Id",
            "type": "string"
          },
          "host_key": {
            "title": "Host Key",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "host_key",
          "host_id",
          "drained_workers",
          "status"
        ],
        "title": "DrainHostResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ImageJobActionResponse": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "job_id": {
            "format": "uuid",
            "title": "Job Id",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "action",
          "status",
          "message"
        ],
        "title": "ImageJobActionResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBasic": {
        "scheme": "basic",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Fictional World API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/admin/v1/gallery": {
      "get": {
        "description": "Return gallery items for the given world, optionally filtered by display_status.",
        "operationId": "list_gallery_admin_v1_gallery_get",
        "parameters": [
          {
            "description": "World UUID",
            "in": "query",
            "name": "world_id",
            "required": true,
            "schema": {
              "description": "World UUID",
              "format": "uuid",
              "title": "World Id",
              "type": "string"
            }
          },
          {
            "description": "Filter by display_status: auto_selected|user_selected|rejected|hidden|superseded",
            "in": "query",
            "name": "display_status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by display_status: auto_selected|user_selected|rejected|hidden|superseded",
              "title": "Display Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GalleryItemRead"
                  },
                  "title": "Response List Gallery Admin V1 Gallery Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "summary": "List gallery items for a world",
        "tags": [
          "admin-stage4"
        ]
      }
    },
    "/admin/v1/hosts/{host_key}/drain": {
      "post": {
        "description": "Mark all workers on ``host_key`` as draining.\n\nNew tasks will not be routed to draining workers.  In-flight tasks complete\nor their leases expire and are re-queued.\n\nHandbook: 29 \u00a7S4-OPS-001; runbook-drain-failover.md \u00a71.",
        "operationId": "drain_host_admin_v1_hosts__host_key__drain_post",
        "parameters": [
          {
            "in": "path",
            "name": "host_key",
            "required": true,
            "schema": {
              "title": "Host Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DrainHostResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "summary": "Drain all workers on a named host",
        "tags": [
          "admin-stage4"
        ]
      }
    },
    "/admin/v1/image-jobs": {
      "get": {
        "description": "Return image jobs for the given world, optionally filtered by status.",
        "operationId": "list_image_jobs_admin_v1_image_jobs_get",
        "parameters": [
          {
            "description": "World UUID to filter image jobs",
            "in": "query",
            "name": "world_id",
            "required": true,
            "schema": {
              "description": "World UUID to filter image jobs",
              "format": "uuid",
              "title": "World Id",
              "type": "string"
            }
          },
          {
            "description": "Filter by job status: queued|running|succeeded|failed|rejected|approved|cancelled|dead_letter",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by job status: queued|running|succeeded|failed|rejected|approved|cancelled|dead_letter",
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ImageJobRead"
                  },
                  "title": "Response List Image Jobs Admin V1 Image Jobs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "summary": "List image jobs for a world",
        "tags": [
          "admin-stage4"
        ]
      }
    },
    "/admin/v1/image-jobs/{job_id}/approve": {
      "post": {
        "description": "Find the gallery item linked to this job and approve it.\n\nDelegates to ``ImageQCService.approve`` on the most-recent gallery item\nfor the job.",
        "operationId": "approve_image_job_admin_v1_image_jobs__job_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageJobActionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "summary": "Manually approve an image job's gallery item",
        "tags": [
          "admin-stage4"
        ]
      }
    },
    "/admin/v1/image-jobs/{job_id}/cancel": {
      "post": {
        "description": "Cancel an image job that has not yet completed.",
        "operationId": "cancel_image_job_admin_v1_image_jobs__job_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageJobActionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "summary": "Cancel a queued or running image job",
        "tags": [
          "admin-stage4"
        ]
      }
    },
    "/admin/v1/image-jobs/{job_id}/reject": {
      "post": {
        "description": "Find the gallery item linked to this job and reject it.",
        "operationId": "reject_image_job_admin_v1_image_jobs__job_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageJobActionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "summary": "Manually reject an image job's gallery item",
        "tags": [
          "admin-stage4"
        ]
      }
    },
    "/admin/v1/image-jobs/{job_id}/retry": {
      "post": {
        "description": "Reset a failed job to 'queued' for regeneration (if attempts allow).\n\nDelegates to ``ImageQCService.mark_regenerate``.",
        "operationId": "retry_image_job_admin_v1_image_jobs__job_id__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageJobActionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "summary": "Retry a failed or dead-letter image job",
        "tags": [
          "admin-stage4"
        ]
      }
    },
...
  }
}